Split the remote ghidra test out
